The first version of the product
The first version of the program, which has the basic functions of parsing the site. Were written functions and fields for the operation of parsing the main page and the results of the query, selecting N number of pages and parsing a specific page with a structure. Implemented a class for storing information about buildings.